tests/wpt/tests/css/css-values/lh-unit-004.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Values and Units Test: lh not affected by &lt;select&gt; fixup</title>
<link rel="help" href="https://drafts.csswg.org/css-values-4/#font-relative-lengths">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  select {
    line-height: 100px;
    width: 5lh;
  }
</style>
<select id="lh"></select>
<script>
  test(() => {
    assert_equals(getComputedStyle(lh).width, "500px");
  }, "lh must be relative to computed line-height before select element fixup");
</script>
